Update wordpress-backup.sh
Modified the script so that the backup folder and the backup file name also contains the site domain name. 

When the sitename is www.domain.com then the filename will start with www_domain and when the sitename is xyz.domain.com then the filename will start with xyz_domain. The date and time will be appended at the end of the filename

In this script, the SITE_URL variable will get the site URL (like www.domain.com or xyz.domain.com). The DOMAIN_NAME variable will then process this URL to create a filename-friendly version (www_domain or xyz_domain). This domain name, along with the current date and time, is used in both the database backup filename and the WordPress directory archive filename.